@@ -30,55 +31,57 @@
     "await-delay": "^1.0.0",
     "babel-polyfill": "^6.23.0",
     "blockies": "0.0.2",
+    "classnames": "^2.2.5",
     "cross-env": "^5.0.5",
     "d3": "^4.9.1",
     "d3-scale-chromatic": "^1.1.1",
     "decimal.js": "^7.2.3",
     "es6-promise": "4.0.5",
-    "lodash": "4.17.2",
+    "lodash": "4.17.4",
     "moment": "^2.17.1",
     "moment-duration-format": "^1.3.0",
     "normalizr": "^3.2.2",
-    "react": "15.4.1",
+    "react": "15.6.1",
     "react-autosize-textarea": "^0.4.7",
     "react-autosuggest": "^9.0.0",
-    "react-custom-scroll": "^1.8.0",
-    "react-datetime": "^2.8.10",
-    "react-dnd": "^2.2.4",
-    "react-dnd-html5-backend": "^2.2.4",
-    "react-dom": "15.4.1",
-    "react-draggable": "^2.2.3",
-    "react-intl": "2.1.2",
+    "react-custom-scroll": "2.0.1",
+    "react-datetime": "^2.10.1",
+    "react-dnd": "2.5.1",
+    "react-dnd-html5-backend": "2.5.1",
+    "react-dom": "15.6.1",
+    "react-draggable": "3.0.3",
+    "react-intl": "2.4.0",
     "react-portal": "^3.0.0",
     "react-redux": "^4.4.6",
     "react-router": "3.0.2",
     "react-router-redux": "4.0.7",
+    "react-select": "^1.0.0-rc.10",
     "react-textarea-autosize": "^4.0.5",
+    "react-transition-group": "^2.2.0",
     "recharts": "^1.0.0-alpha.1",
     "redux": "^3.6.0",
     "redux-actions": "^2.0.1",
-    "redux-form": "^7.0.3",
+    "redux-form": "7.0.4",
     "redux-thunk": "^2.2.0",
     "reselect": "^2.5.4",
     "sha1": "^1.1.1",
+    "truffle-contract": "^3.0.0",
     "uuid": "^3.1.0",
-    "web3": "0.18.4",
-    "truffle-contract": "3.0.0"
+    "web3": "0.18.4"
   },
   "devDependencies": {
-    "autoprefixer": "^7.1.1",
-    "babel-core": "^6.23.1",
+    "autoprefixer": "7.1.4",
+    "babel-core": "6.26.0",
     "babel-eslint": "7.1.1",
-    "babel-loader": "6.2.9",
-    "babel-plugin-react-intl": "2.2.0",
+    "babel-plugin-lodash": "^3.2.11",
     "babel-plugin-transform-async-to-generator": "^6.24.1",
+    "babel-loader": "7.1.2",
+    "babel-plugin-react-intl": "2.3.1",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-plugin-transform-function-bind": "^6.22.0",
-    "babel-plugin-transform-react-remove-prop-types": "^0.4.8",
-    "babel-preset-es2015": "6.14.0",
-    "babel-preset-es2016": "^6.22.0",
-    "babel-preset-react": "6.11.1",
-    "babel-preset-stage-0": "6.5.0",
+    "babel-preset-env": "1.6.0",
+    "babel-preset-react": "6.24.1",
+    "babel-preset-stage-0": "6.24.1",
     "bootstrap-loader": "^2.1.0",
     "bootstrap-sass": "^3.3.7",
     "css-loader": "^0.27.3",
@@ -108,7 +111,7 @@
     "style-loader": "^0.13.2",
     "truffle-hdwallet-provider": "0.0.3",
     "url-loader": "^0.5.9",
-    "webpack": "^2.2.1",
-    "webpack-dev-server": "^2.4.1"
+    "webpack": "3.6.0",
+    "webpack-dev-server": "2.8.2"
   }
 }
