@@ -3,8 +3,7 @@
 import { mapValues, startsWith, isArray } from 'lodash'
 import Decimal from 'decimal.js'
 import { HEX_VALUE_REGEX, OUTCOME_TYPES } from 'utils/constants'
-
-const config = require('config.json')
+import config from 'config.json'
 
 export const hexWithoutPrefix = (value) => {
   if (HEX_VALUE_REGEX.test(value)) {
