@@ -14,14 +14,32 @@ const Header = ({ version, defaultAccount, currentBalance, currentProvider, getP
           <div className="headerLogo" />
         </Link>
       </div>
-      <div className="headerContainer__group headerContainer__group--left headerContainer__group--version">{version}</div>
-      <div className="headerContainer__group headerContainer__group--left">
-        <Link to="/dashboard" activeClassName="headerContainer__navLink--active" className="headerContainer__navLink">Dashboard</Link>
-        <Link to="/markets/list" activeClassName="headerContainer__navLink--active" className="headerContainer__navLink">Markets</Link>
-        {defaultAccount && <Link to="/transactions" activeClassName="headerContainer__navLink--active" className="headerContainer__navLink">Transactions</Link>}
+      <div className="headerContainer__group headerContainer__group--left headerContainer__group--version">
+        {version}
       </div>
-      <div className="headerContainer__group headerContainer__group--right">
-        {(defaultAccount && currentProvider) && getProviderIcon(currentProvider)}
+      <div className="headerContainer__group headerContainer__group--left">
+        <Link to="/dashboard" activeClassName="headerContainer__navLink--active" className="headerContainer__navLink">
+          Dashboard
+        </Link>
+        <Link
+          to="/markets/list"
+          activeClassName="headerContainer__navLink--active"
+          className="headerContainer__navLink"
+        >
+          Markets
+        </Link>
+        {defaultAccount && (
+          <Link
+            to="/transactions"
+            activeClassName="headerContainer__navLink--active"
+            className="headerContainer__navLink"
+          >
+            Transactions
+          </Link>
+        )}
+      </div>
+      <div className="headerContainer__group headerContainer__group--right account">
+        {defaultAccount && currentProvider && getProviderIcon(currentProvider)}
         <div className="headerContainer__account">
           <DecimalValue value={currentBalance} className="headerContainer__account--text" />&nbsp;<span className="headerContainer__account--text">ETH</span>
           <Identicon className="" />
