@@ -25,48 +25,44 @@ const OutcomeCategorical = ({ market, opts = {} }) => {
     const tokenDistributionInt = tokenDistribution.map(outcome => parseInt(parseFloat(outcome) * 10000, 10))
     const trendingOutcomeIndex = tokenDistributionInt.indexOf(Math.max(...tokenDistributionInt))
     return (
-      <div className="outcomes outcomes--categorical row">
-        <div className="col-md-6">
-          <div className="row">
-            <div className="col-md-6">
-              <div className={'entry__color pull-left'} style={{ backgroundColor: COLOR_SCHEME_DEFAULT[trendingOutcomeIndex] }} />
-              <div className="pull-left outcomes--categorical--top-6">{ renderOutcomes[trendingOutcomeIndex] }</div>
-            </div>
-            <div className="col-md-2 outcomes--top-6">
-              {market.marginalPrices ? Math.round(market.marginalPrices[trendingOutcomeIndex] * 100).toFixed(0) : 0}%
-            </div>
-          </div>
+      <div className="outcomes outcomes--categorical">
+        <div className="entry__color" style={{ backgroundColor: COLOR_SCHEME_DEFAULT[trendingOutcomeIndex] }} />
+        <div className="outcome">{ renderOutcomes[trendingOutcomeIndex] }</div>
+        <div>
+          {market.marginalPrices ? Math.round(market.marginalPrices[trendingOutcomeIndex] * 100).toFixed(0) : 0}%
         </div>
-        <div className="col-md-4 outcomes--top-6">
+        <div>
           { showDate ? moment(market.eventDescription.resolutionDate).format(dateFormat) : ''}
         </div>
       </div>
     )
   }
 
-  return (<div className={`${className} outcomes outcomes--categorical`}>
-    {renderOutcomes.map((outcome, outcomeIndex) => {
-      if (market.oracle.isOutcomeSet && market.oracle.outcome !== outcomeIndex) {
-        return <div key={outcomeIndex} className="outcome" />
-      }
+  return (
+    <div className={`${className} outcomes outcomes--categorical`}>
+      {renderOutcomes.map((outcome, outcomeIndex) => {
+        if (market.oracle.isOutcomeSet && market.oracle.outcome !== outcomeIndex) {
+          return <div key={outcomeIndex} className="outcome" />
+        }
 
-      return (
-        <div key={outcomeIndex} className="outcome">
-          <div className="outcome__bar">
-            <div
-              className="outcome__bar--inner"
-              style={{ width: `${tokenDistribution[outcomeIndex] * 100}%`, backgroundColor: COLOR_SCHEME_DEFAULT[outcomeIndex] }}
-            >
-              <div className="outcome__bar--label">
-                { renderOutcomes[outcomeIndex] }
-                <div className="outcome__bar--value">{ `${Math.round(tokenDistribution[outcomeIndex] * 100).toFixed(0)}%` }</div>
+        return (
+          <div key={outcomeIndex} className="outcome">
+            <div className="outcome__bar">
+              <div
+                className="outcome__bar--inner"
+                style={{ width: `${tokenDistribution[outcomeIndex] * 100}%`, backgroundColor: COLOR_SCHEME_DEFAULT[outcomeIndex] }}
+              >
+                <div className="outcome__bar--label">
+                  { renderOutcomes[outcomeIndex] }
+                  <div className="outcome__bar--value">{ `${Math.round(tokenDistribution[outcomeIndex] * 100).toFixed(0)}%` }</div>
+                </div>
               </div>
             </div>
           </div>
-        </div>
-      )
-    })}
-  </div>)
+        )
+      })}
+    </div>
+  )
 }
 
 OutcomeCategorical.propTypes = {
