@@ -1,4 +1,4 @@
-import React, { Component } from 'react'
+import { Component } from 'react'
 import PropTypes from 'prop-types'
 import { connectBlockchain, initGnosis } from 'actions/blockchain'
 import { getSelectedProvider } from 'selectors/blockchain'
