@@ -3,6 +3,8 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
+import TransitionGroup from 'react-transition-group/TransitionGroup'
+import CSSTransition from 'react-transition-group/CSSTransition'
 
 import { connectBlockchain } from 'actions/blockchain'
 
@@ -27,20 +29,28 @@ class App extends Component {
       )
     }
 
+    const currentKey = this.props.location.pathname.split('/')[2] || this.props.location.pathname.split('/')[1] || '/'
+    const timeout = { enter: 200, exit: 200 }
+
     return (
       <div className="appContainer">
         <HeaderContainer version={process.env.VERSION} />
         {this.props.hasWallet && <TransactionFloaterContainer />}
-        {this.props.children}
+        <TransitionGroup>
+          <CSSTransition key={currentKey} classNames="page-transition" timeout={timeout}>
+            {this.props.children}
+          </CSSTransition>
+        </TransitionGroup>
       </div>
     )
   }
 }
 
 App.propTypes = {
+  blockchainConnection: PropTypes.bool,
   children: PropTypes.node,
   connectBlockchain: PropTypes.func,
-  blockchainConnection: PropTypes.bool,
+  location: PropTypes.object,
   hasWallet: PropTypes.bool,
 }
 
