@@ -1,4 +1,4 @@
-import { hashHistory } from 'react-router'
+import { browserHistory } from 'react-router'
 import { createStore, applyMiddleware, compose } from 'redux'
 import thunk from 'redux-thunk'
 import { routerMiddleware } from 'react-router-redux'
@@ -10,14 +10,18 @@ import Notifications from 'middlewares/Notifications'
 
 import reducer from 'reducers'
 
+const middlewares = [
+  thunk,
+  routerMiddleware(browserHistory),
+  Notifications,
+  LocalStorageLoad,
+  LocalStorageDump,
+  CrashReporter,
+]
+
 const enhancers = [
   applyMiddleware(
-    thunk,
-    routerMiddleware(hashHistory),
-    Notifications,
-    LocalStorageLoad,
-    LocalStorageDump,
-    CrashReporter,
+    ...middlewares,
   ),
 ]
 
@@ -26,8 +30,6 @@ if (window.devToolsExtension) {
   enhancers.push(window.devToolsExtension())
 }
 
-const store = createStore(
-  reducer, compose(...enhancers),
-)
+const store = createStore(reducer, compose(...enhancers))
 
 export default store
